+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/API/Types.hs
@@ -43,17 +43,25 @@ import Cardano.Ledger.Hashes as X (
   ScriptHash (..),
  )
 import Cardano.Ledger.Keys as X (
+  CertifiedVRF,
   GenDelegPair (..),
   GenDelegs (..),
   Hash,
+  KESignable,
   KeyHash (..),
   KeyRole (..),
   SignKeyDSIGN,
+  SignKeyKES,
+  SignKeyVRF,
   SignedDSIGN,
+  SignedKES,
   VKey (..),
+  VerKeyKES,
+  VerKeyVRF,
   WitVKey (..),
   coerceKeyRole,
   hashKey,
+  hashVerKeyVRF,
  )
 import Cardano.Ledger.Keys.Bootstrap as X (
   BootstrapWitness (..),