+++ libs/cardano-protocol-tpraos/src/Cardano/Protocol/TPraos/Rules/Overlay.hs
@@ -22,6 +22,7 @@ module Cardano.Protocol.TPraos.Rules.Overlay (
 )
 where
 
+import qualified Cardano.Crypto.KES as KES
 import qualified Cardano.Crypto.VRF as VRF
 import Cardano.Ledger.BHeaderView (isOverlaySlot)
 import Cardano.Ledger.BaseTypes (
@@ -48,8 +49,6 @@ import Cardano.Ledger.Keys (
   GenDelegs (..),
   KeyHash (..),
   KeyRole (..),
-  KeyRoleVRF (..),
-  VRFVerKeyHash (..),
   coerceKeyRole,
   hashKey,
  )
@@ -134,7 +133,7 @@ data OverlayPredicateFailure c
 
 instance
   ( Crypto c
-  , KESignable c (BHBody c)
+  , KES.Signable (KES c) (BHBody c)
   , VRF.Signable (VRF c) Seed
   ) =>
   STS (OVERLAY c)
@@ -234,7 +233,7 @@ pbftVrfChecks genDelegVRFVerKeyHash eta0 bhb = do
 overlayTransition ::
   forall c.
   ( Crypto c
-  , KESignable c (BHBody c)
+  , KES.Signable (KES c) (BHBody c)
   , VRF.Signable (VRF c) Seed
   ) =>
   TransitionRule (OVERLAY c)
@@ -283,7 +282,7 @@ instance
 
 instance
   ( Crypto c
-  , KESignable c (BHBody c)
+  , KES.Signable (KES c) (BHBody c)
   , VRF.Signable (VRF c) Seed
   ) =>
   Embed (OCERT c) (OVERLAY c)