+++ libs/cardano-protocol-tpraos/testlib/Test/Cardano/Protocol/Crypto/KES.hs
@@ -12,7 +12,7 @@ import Cardano.Protocol.Crypto
 
 data KESKeyPair c = KESKeyPair
   { kesSignKey :: !(KES.UnsoundPureSignKeyKES (KES c))
-  , kesVerKey :: !(VerKeyKES c)
+  , kesVerKey :: !(KES.VerKeyKES (KES c))
   }
 
 instance Show (KES.VerKeyKES (KES c)) => Show (KESKeyPair c) where